broke search page into smaller react components
