+++ frontend/src/assets/styles/_makeover.css
@@ -43,6 +43,22 @@
   --tz-dropzone-hover-bg: rgba(27, 34, 48, 0.78);
   --tz-upload-panel-bg: rgba(23, 31, 51, 0.94);
   --tz-lightbox-bg: rgba(8, 11, 20, 0.86);
+  --tz-ui-page: var(--tz-app-bg);
+  --tz-ui-surface: rgba(15, 23, 42, 0.78);
+  --tz-ui-surface-raised: rgba(23, 31, 51, 0.96);
+  --tz-ui-surface-subtle: rgba(30, 41, 59, 0.72);
+  --tz-ui-control-bg: rgba(30, 41, 59, 0.96);
+  --tz-ui-control-hover-bg: rgba(51, 65, 85, 0.98);
+  --tz-ui-text: #f8fafc;
+  --tz-ui-text-soft: #e2e8f0;
+  --tz-ui-text-muted: #cbd5e1;
+  --tz-ui-text-faint: #94a3b8;
+  --tz-ui-border: rgba(226, 232, 240, 0.22);
+  --tz-ui-border-strong: rgba(226, 232, 240, 0.38);
+  --tz-ui-accent: var(--tz-turmeric);
+  --tz-ui-accent-contrast: #171101;
+  --tz-ui-focus-ring: rgba(224, 168, 46, 0.34);
+  --tz-ui-shadow-panel: 0 18px 48px rgba(0, 0, 0, 0.28);
   --reader-page-bg: linear-gradient(135deg, #0a0a0a 0%, #1a1a1a 50%, #0f0f0f 100%);
   --reader-title-gradient: linear-gradient(135deg, #fff 0%, #aaa 100%);
   --reader-muted: #888;
@@ -111,6 +127,22 @@
   --tz-dropzone-hover-bg: rgba(241, 245, 249, 0.96);
   --tz-upload-panel-bg: rgba(255, 255, 255, 0.90);
   --tz-lightbox-bg: rgba(17, 24, 39, 0.54);
+  --tz-ui-page: var(--tz-app-bg);
+  --tz-ui-surface: rgba(255, 255, 255, 0.82);
+  --tz-ui-surface-raised: rgba(255, 255, 255, 0.96);
+  --tz-ui-surface-subtle: rgba(241, 245, 249, 0.86);
+  --tz-ui-control-bg: rgba(255, 255, 255, 0.96);
+  --tz-ui-control-hover-bg: #edf2f8;
+  --tz-ui-text: #0f172a;
+  --tz-ui-text-soft: #1f2937;
+  --tz-ui-text-muted: #475569;
+  --tz-ui-text-faint: #64748b;
+  --tz-ui-border: rgba(15, 23, 42, 0.14);
+  --tz-ui-border-strong: rgba(15, 23, 42, 0.24);
+  --tz-ui-accent: var(--tz-turmeric);
+  --tz-ui-accent-contrast: #171101;
+  --tz-ui-focus-ring: rgba(183, 121, 31, 0.28);
+  --tz-ui-shadow-panel: 0 18px 44px rgba(31, 41, 55, 0.12);
   --reader-page-bg: linear-gradient(135deg, #f8fafc 0%, #eef4fb 50%, #ffffff 100%);
   --reader-title-gradient: linear-gradient(135deg, #111827 0%, #365ea8 100%);
   --reader-muted: #4b5563;
