+++ frontend/src/pages/FabricSearch.tsx
@@ -1,956 +1,21 @@
-import {  useCallback, useEffect, useId, useMemo, useRef, useState } from "react";
-import { BASE_URL, FULL_API_URL } from "../constants";
+import { useCallback, useEffect, useId, useMemo, useRef, useState } from "react";
+import type { ChangeEvent, MouseEventHandler, WheelEventHandler } from "react";
 
 import Loader from "../components/Loader";
 import Notification from "../components/Notification";
+import CropDrawer from "../components/search/CropDrawer";
+import ImagePreview from "../components/search/ImagePreview";
+import Lightbox from "../components/search/Lightbox";
+import ResultsSection from "../components/search/ResultsSection";
+import SettingsPanel from "../components/search/SettingsPanel";
+import SearchHero from "../components/search/SearchHero";
+import StickySearchBar from "../components/search/StickySearchBar";
+import { MIN_CROP_SIZE } from "../components/search/searchConfig";
+import { clampCropRectToBounds } from "../components/search/searchUtils";
+import type { CropRect, NotificationState } from "../components/search/types";
+import { useSearch } from "../components/search/useSearch";
 import { throttle } from "../utils/throttle";
 import "@/assets/styles/FabricSearch.css";
-
-// ─── Types ───────────────────────────────────────────────────────────────────
-
-type NotificationState = { message: string; type: "success" | "error" } | null;
-type DbOp = "create" | "update" | null;
-
-interface ResultItem {
-  imageSrc: string;
-  filename: string;
-  audioSrc?: string;
-}
-
-interface SearchApiResponse {
-  message: string;
-  results: string[];
-  pagination: {
-    page: number;
-    per_page: number;
-    total_results: number;
-    total_pages: number;
-    has_next: boolean;
-    has_prev: boolean;
-  };
-}
-
-type CropRect = { x: number; y: number; w: number; h: number };
-
-const MIN_CROP_SIZE = 40;
-
-function clampCropRectToBounds(rect: CropRect, imgWidth: number, imgHeight: number): CropRect {
-  if (imgWidth <= 0 || imgHeight <= 0) return rect;
-
-  const maxW = Math.max(MIN_CROP_SIZE, imgWidth);
-  const maxH = Math.max(MIN_CROP_SIZE, imgHeight);
-  const w = Math.min(Math.max(MIN_CROP_SIZE, rect.w), maxW);
-  const h = Math.min(Math.max(MIN_CROP_SIZE, rect.h), maxH);
-  const x = Math.min(Math.max(0, rect.x), Math.max(0, imgWidth - w));
-  const y = Math.min(Math.max(0, rect.y), Math.max(0, imgHeight - h));
-
-  return {
-    x: Math.round(x),
-    y: Math.round(y),
-    w: Math.round(w),
-    h: Math.round(h),
-  };
-}
-
-// ─── Constants ────────────────────────────────────────────────────────────────
-
-const CATEGORIES = [
-  { id: "stock", label: "Stock", icon: "📦" },
-  { id: "fabric", label: "Fabric", icon: "🧵" },
-  { id: "design", label: "Design", icon: "🎨" },
-  { id: "product", label: "Product", icon: "🖼️" },
-];
-
-const API_BASE = FULL_API_URL;
-const ASSET_BASE = BASE_URL;
-const CDN_BASE = (import.meta.env.VITE_AWS_PUBLIC_URL ?? "https://cdn.threadzip.com").replace(/\/$/, "");
-const USER_FRIENDLY_SERVER_ERROR = "Unable to connect to the server, please try after some time.";
-
-// ─── Helpers ──────────────────────────────────────────────────────────────────
-
-function toCdnUrl(src: string | undefined): string {
-  if (!src) return "";
-
-  const clean = String(src).trim();
-
-  if (/^(https?:|blob:|data:)/i.test(clean)) {
-    return clean;
-  }
-
-  if (clean.startsWith("/api/") || clean.startsWith("/assets/")) {
-    return `${ASSET_BASE}${clean}`;
-  }
-
-  if (clean.startsWith("images/")) {
-    return `${CDN_BASE}/${clean}`;
-  }
-
-  return `${CDN_BASE}/images/${clean.replace(/^\/+/, "")}`;
-}
-
-function toResultItem(raw: string): ResultItem {
-  const filename = raw.split("/").pop() ?? raw;
-  return { imageSrc: raw, filename };
-}
-
-function cleanName(filename: string): string {
-  return filename ? filename.split("_")[0].split(".")[0] : "";
-}
-
-// ─── useSearch hook ───────────────────────────────────────────────────────────
-
-function useSearch() {
-  const [loading, setLoading] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-  const [results, setResults] = useState<ResultItem[]>([]);
-
-  const runImageSearch = useCallback(async (file: File, category?: string[], limit = 20, preserveResultsOnError = false) => {
-    setLoading(true);
-    setError(null);
-    try {
-      const form = new FormData();
-      form.append("file", file);
-      form.append("limit", String(limit));
-      form.append("page", "1");
-      form.append("per_page", String(limit));
-      if (category?.length) category.forEach((c) => { form.append("category", c) });
-      const res = await fetch(`${API_BASE}/search`, { method: "POST", body: form });
-      if (!res.ok) {
-        const err = await res.json().catch(() => ({}));
-        throw new Error(err?.detail ?? `Search failed (${res.status})`);
-      }
-      const data: SearchApiResponse = await res.json();
-      setResults((data.results ?? []).map(toResultItem));
-    } catch (e) {
-      console.error("Image search failed:", e);
-      setError(USER_FRIENDLY_SERVER_ERROR);
-      if (!preserveResultsOnError) setResults([]);
-    } finally {
-      setLoading(false);
-    }
-  }, []);
-
-  const runTextSearch = useCallback(async (term: string, category?: string[], limit = 20, preserveResultsOnError = false) => {
-    setLoading(true);
-    setError(null);
-    try {
-      const form = new FormData();
-      form.append("search_term", term);
-      form.append("limit", String(limit));
-      form.append("page", "1");
-      form.append("per_page", String(limit));
-      if (category?.length) category.forEach((c) => { form.append("category", c) });
-      const res = await fetch(`${API_BASE}/search`, { method: "POST", body: form });
-      if (!res.ok) {
-        const err = await res.json().catch(() => ({}));
-        throw new Error(err?.detail ?? `Search failed (${res.status})`);
-      }
-      const data: SearchApiResponse = await res.json();
-      setResults((data.results ?? []).map(toResultItem));
-    } catch (e) {
-      console.error("Text search failed:", e);
-      setError(USER_FRIENDLY_SERVER_ERROR);
-      if (!preserveResultsOnError) setResults([]);
-    } finally {
-      setLoading(false);
-    }
-  }, []);
-
-  const clear = useCallback(() => {
-    setResults([]);
-    setError(null);
-  }, []);
-
-  return { loading, error, results, runImageSearch, runTextSearch, clear };
-}
-
-// ─── DB endpoint ──────────────────────────────────────────────────────────────
-
-async function callDbEndpoint(op: "create" | "update"): Promise<string> {
-  const url =
-    op === "create"
-      ? `${API_BASE}/database/create/table`
-      : `${API_BASE}/database/update/table`;
-  const res = await fetch(url, { method: "PUT" });
-  const data = await res.json().catch(() => ({}));
-  if (!res.ok) throw new Error(data?.detail ?? `Request failed (${res.status})`);
-  return data?.message ?? "Done.";
-}
-
-// ─── CategoryPicker ───────────────────────────────────────────────────────────
-
-interface CategoryPickerProps {
-  selected: string[];
-  onChange: (cats: string[]) => void;
-  compact?: boolean;
-}
-
-function CategoryPicker({ selected, onChange, compact = false }: CategoryPickerProps) {
-  const [tempSelected, setTempSelected] = useState(selected);
-  const toggle = (id: string) => {
-    const next = tempSelected.includes(id)
-      ? tempSelected.filter((c) => c !== id)
-      : [...tempSelected, id];
-    setTempSelected(next);
-    // Non-compact (Hero / ImagePreview): propagate immediately so the parent's
-    // selectedCategories stays in sync. Compact mode defers to the Apply button.
-    if (!compact) {
-      onChange(next);
-    }
-  };
-  const allOn = tempSelected.length === CATEGORIES.length;
-  const toggleAll = () => {
-    const next = allOn ? [] : CATEGORIES.map((c) => c.id);
-    setTempSelected(next);
-    if (!compact) {
-      onChange(next);
-    }
-  };
-
-  const applySearch = () => {
-    onChange(tempSelected);
-  };
-
-  useEffect(() => {
-    setTempSelected(selected);
-  }, [selected]);
-
-  return (
-    <div className={`category-picker${compact ? " category-picker--compact" : ""}`}>
-      <div className="category-picker__header">
-        <span className="category-picker__title">Filter by Category</span>
-        <button className="category-picker__toggle-all" onClick={toggleAll} type="button">
-          {allOn ? "Clear all" : "Select all"}
-        </button>
-      </div>
-
-      <div className="category-picker__grid">
-        {CATEGORIES.map((cat) => {
-          const active = tempSelected.includes(cat.id);
-          return (
-            <button
-              key={cat.id}
-              className={`category-picker__chip${active ? " category-picker__chip--active" : ""}`}
-              onClick={() => toggle(cat.id)}
-              type="button"
-            >
-              {active && <span className="category-picker__chip-check">✓</span>}
-              <span className="category-picker__chip-icon">{cat.icon}</span>
-              <span className="category-picker__chip-label">{cat.label}</span>
-            </button>
-          );
-        })}
-        {compact && (
-          <button className="btn btn-primary" onClick={()=> applySearch()} type="button">
-            {'Apply filter →'}
-          </button>
-        )}
-      </div>
-    </div>
-  );
-}
-
-// ─── DbControlPanel ───────────────────────────────────────────────────────────
-
-function DbControlPanel() {
-  const [activeOp, setActiveOp] = useState<DbOp>(null);
-  const [notification, setNotification] = useState<NotificationState>(null);
-
-  const handleOp = async (op: "create" | "update") => {
-    if (activeOp) return;
-    setActiveOp(op);
-    setNotification(null);
-    try {
-      const msg = await callDbEndpoint(op);
-      setNotification({ message: msg, type: "success" });
-    } catch (e: unknown) {
-      console.error("Database operation failed:", e);
-      setNotification({ message: USER_FRIENDLY_SERVER_ERROR, type: "error" });
-    } finally {
-      setActiveOp(null);
-    }
-  };
-
-  return (
-    <div className="db-panel">
-      <div className="db-panel__header">
-        <span className="db-panel__pulse" />
-        <span className="db-panel__label">Vector Database</span>
-      </div>
-
-      <div className="db-panel__actions">
-        <button
-          className={`db-panel__btn db-panel__btn--create${activeOp === "create" ? " db-panel__btn--loading" : ""}`}
-          onClick={() => handleOp("create")}
-          disabled={!!activeOp}
-          title="Rebuild vector table from scratch"
-        >
-          {activeOp === "create"
-            ? <span className="db-panel__spinner" />
-            : <span className="db-panel__btn-icon">⬡</span>}
-          Create Table
-        </button>
-
-        <button
-          className={`db-panel__btn db-panel__btn--update${activeOp === "update" ? " db-panel__btn--loading" : ""}`}
-          onClick={() => handleOp("update")}
-          disabled={!!activeOp}
-          title="Add new images to existing table"
-        >
-          {activeOp === "update"
-            ? <span className="db-panel__spinner" />
-            : <span className="db-panel__btn-icon">↻</span>}
-          Update Table
-        </button>
-      </div>
-
-      {notification && (
-        <div className={`db-panel__notification db-panel__notification--${notification.type}`}>
-          {notification.message}
-        </div>
-      )}
-    </div>
-  );
-}
-
-// ─── SettingsPanel ────────────────────────────────────────────────────────────
-
-function SettingsPanel() {
-  const [open, setOpen] = useState(false);
-
-  return (
-    <div className="settings-panel">
-      <button
-        className="settings-panel__toggle"
-        onClick={() => setOpen((v) => !v)}
-        type="button"
-        aria-expanded={open}
-        title="Settings"
-      >
-        ⚙️ &nbsp;Settings
-        <span className="settings-panel__chevron">{open ? "▲" : "▼"}</span>
-      </button>
-
-      {open && (
-        <div className="settings-panel__content">
-          <DbControlPanel />
-        </div>
-      )}
-    </div>
-  );
-}
-
-// ─── StickySearchBar ──────────────────────────────────────────────────────────
-
-interface StickySearchBarProps {
-  textQuery: string;
-  setTextQuery: (v: string) => void;
-  onSearch: () => void;
-  onClear: () => void;
-  loading: boolean;
-  isImageMode: boolean;
-  previewUrl: string | null;
-  onRecrop: () => void;
-  fileInputId: string;
-  onFileChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
-  selectedCategories: string[];
-  onSetCategories: (cats: string[]) => void;
-  searchLimit: number;
-  onSetLimit: (v: number) => void;
-}
-
-function StickySearchBar({
-  textQuery,
-  setTextQuery,
-  onSearch,
-  onClear,
-  loading,
-  isImageMode,
-  previewUrl,
-  onRecrop,
-  fileInputId,
-  onFileChange,
-
-  selectedCategories,
-  onSetCategories,
-  searchLimit,
-  onSetLimit,
-}: StickySearchBarProps) {
-  return (
-    <div className="search-bar search-bar--sticky">
-      <div className="search-bar__inner">
-
-        {isImageMode && previewUrl ? (
-          <div className="search-bar__image-row">
-            <div className="search-bar__thumb-wrap">
-              <img src={previewUrl} alt="query" className="search-bar__thumb" />
-            </div>
-
-            <div className="search-bar__image-actions">
-              <button className="btn btn--outline btn--sm" onClick={onRecrop}>
-                ✂️ Recrop
-              </button>
-
-              <input
-                id={fileInputId}
-                type="file"
-                accept="image/*"
-                onChange={onFileChange}
-                className="search-bar__file-input"
-              />
-
-              <button
-                className="btn btn--ghost btn--sm"
-                onClick={() => document.getElementById(fileInputId)?.click()}
-              >
-                📷 New Image
-              </button>
-            </div>
-          </div>
-        ) : (
-          <div className="search-bar__text-row">
-            <div className="search-bar__input-wrap">
-              <input
-                type="text"
-                className="search-bar__input"
-                placeholder="Refine search…"
-                value={textQuery}
-                onChange={(e) => setTextQuery(e.target.value)}
-                onKeyDown={(e) => e.key === "Enter" && onSearch()}
-              />
-            </div>
-
-            <button
-              className="btn btn--primary btn--sm"
-              onClick={onSearch}
-              disabled={loading || !textQuery.trim()}
-            >
-              Search
-            </button>
-          </div>
-        )}
-
-        <button className="btn btn--ghost btn--sm" onClick={onClear} type="button">
-          <span aria-hidden="true">✕</span>
-          <span>Clear</span>
-        </button>
-      </div>
-
-      {/* ✅ NEW: filters always visible */}
-      <div className="search-bar__filters">
-        <CategoryPicker
-          selected={selectedCategories}
-          onChange={onSetCategories}
-          compact
-        />
-
-        <LimitSlider
-          value={searchLimit}
-          onChange={onSetLimit}
-          label="Limit"
-        />
-      </div>
-    </div>
-  );
-}
-// ─── Pagination ───────────────────────────────────────────────────────────────
-
-interface PaginationProps {
-  page: number;
-  totalPages: number;
-  onPrev: () => void;
-  onNext: () => void;
-  variant?: "inline" | "bottom";
-}
-
-function Pagination({ page, totalPages, onPrev, onNext, variant = "bottom" }: PaginationProps) {
-  return (
-    <div className={`pagination pagination--${variant}`}>
-      <button className="pagination__btn" onClick={onPrev} disabled={page === 1}>
-        ← Prev
-      </button>
-      <span className="pagination__label">Page {page} / {totalPages}</span>
-      <button className="pagination__btn" onClick={onNext} disabled={page === totalPages}>
-        Next →
-      </button>
-    </div>
-  );
-}
-
-// ─── ResultCard ───────────────────────────────────────────────────────────────
-
-interface ResultCardProps {
-  item: ResultItem;
-  index: number;
-  onZoom: (src: string, caption: string) => void;
-  onBadImage: (src: string) => void;
-}
-
-function ResultCard({ item, index, onZoom, onBadImage }: ResultCardProps) {
-  const handleClick = () => {
-    if (!item.imageSrc) return;
-    onZoom(toCdnUrl(item.imageSrc), cleanName(item.filename));
-  };
-
-  return (
-    <article
-      className="result-card"
-      style={{ animationDelay: `${index * 40}ms` }}
-    >
-      <div className="result-card__thumb">
-        <img
-          src={toCdnUrl(item.imageSrc)}
-          alt={item.filename}
-          loading="lazy"
-          decoding="async"
-          onError={() => onBadImage(item.imageSrc)}
-          onClick={handleClick}
-        />
-        <button
-          type="button"
-          className="result-card__zoom-btn"
-          aria-label="Zoom"
-          onClick={(e) => { e.stopPropagation(); handleClick(); }}
-        >
-          🔍
-        </button>
-      </div>
-
-      <div className="result-card__name" onClick={handleClick}>
-        {cleanName(item.filename)}
-      </div>
-
-      {item.audioSrc && (
-        <div className="result-card__audio">
-          <audio controls src={item.audioSrc} preload="metadata" controlsList="nodownload" />
-        </div>
-      )}
-    </article>
-  );
-}
-
-// ─── ResultsSection ───────────────────────────────────────────────────────────
-
-interface ResultsSectionProps {
-  results: ResultItem[];
-  paginatedResults: ResultItem[];
-  page: number;
-  totalPages: number;
-  selectedCategories: string[];
-  isTextSearch: boolean;
-  onSetCategories: (cats: string[]) => void;
-  onPrev: () => void;
-  onNext: () => void;
-  onZoom: (src: string, caption: string) => void;
-  onBadImage: (src: string) => void;
-}
-
-function ResultsSection({
-  results,
-  paginatedResults,
-  page,
-  totalPages,
-  onPrev,
-  onNext,
-  onZoom,
-  onBadImage,
-}: ResultsSectionProps) {
-  return (
-    <div className="results-section">
-
-      <div className="results-section__meta">
-        <div>{results.length} results</div>
-        <Pagination page={page} totalPages={totalPages} onPrev={onPrev} onNext={onNext} />
-      </div>
-
-      {/* ✅ ALWAYS visible */}
-      {/* <div className="results-section__filters one">
-        <CategoryPicker selected={selectedCategories} onChange={onSetCategories} compact />
-      </div> */}
-
-      <div className="result-grid result-grid--full">
-        {paginatedResults.map((item, idx: number) => (
-          <ResultCard
-            key={idx}
-            item={item}
-            index={idx}
-            onZoom={onZoom}
-            onBadImage={onBadImage}
-          />
-        ))}
-      </div>
-    </div>
-  );
-}
-
-// ─── LimitSlider ──────────────────────────────────────────────────────────────
-
-interface LimitSliderProps {
-  value: number;
-  onChange: (v: number) => void;
-  label?: string;
-}
-
-function LimitSlider({ value, onChange, label = "Results" }: LimitSliderProps) {
-  return (
-    <div className="limit-slider">
-      <span className="limit-slider__label">{label}</span>
-      <div className="limit-slider__track">
-        <div
-          className="limit-slider__fill"
-          style={{ width: `${((value - 5) / 95) * 100}%` }}
-        />
-        <input
-          type="range"
-          className="limit-slider__input"
-          min={5} max={100} step={5}
-          value={value}
-          onChange={(e) => onChange(Number(e.target.value))}
-        />
-      </div>
-      <span className="limit-slider__value">{value}</span>
-    </div>
-  );
-}
-
-// ─── Lightbox ─────────────────────────────────────────────────────────────────
-
-interface LightboxProps {
-  src: string;
-  caption: string | null;
-  scale: number;
-  offset: { x: number; y: number };
-  onClose: () => void;
-  onWheel: React.WheelEventHandler<HTMLDivElement>;
-  onMouseDown: React.MouseEventHandler<HTMLDivElement>;
-  onMouseMove: React.MouseEventHandler<HTMLDivElement>;
-  onMouseUp: React.MouseEventHandler<HTMLDivElement>;
-  onZoomIn: () => void;
-  onZoomOut: () => void;
-  onReset: () => void;
-}
-
-function Lightbox({
-  src, caption, scale, offset,
-  onClose, onWheel, onMouseDown, onMouseMove, onMouseUp,
-  onZoomIn, onZoomOut, onReset,
-}: LightboxProps) {
-  return (
-    <div
-      className="lightbox"
-      role="dialog"
-      aria-modal="true"
-      aria-label={caption ? `Preview: ${caption}` : "Image preview"}
-      onClick={(e) => {
-        if ((e.target as HTMLElement).classList.contains("lightbox")) onClose();
-      }}
-    >
-      <div
-        className="lightbox__stage"
-        onWheel={onWheel}
-        onMouseDown={onMouseDown}
-        onMouseMove={onMouseMove}
-        onMouseUp={onMouseUp}
-        onMouseLeave={onMouseUp}
-      >
-        <img
-          src={src}
-          alt={caption ?? "preview"}
-          className="lightbox__image"
-          style={{ transform: `translate(${offset.x}px, ${offset.y}px) scale(${scale})` }}
-          draggable={false}
-        />
-
-        {caption && <div className="lightbox__caption">{caption}</div>}
-
-        <div
-          className="lightbox__controls"
-          onMouseDown={(e) => e.stopPropagation()}
-          onClick={(e) => e.stopPropagation()}
-        >
-          <button
-            className="lightbox__control-btn"
-            onClick={onZoomOut}
-            type="button"
-            aria-label="Zoom out"
-          >
-            −
-          </button>
-          <button
-            className="lightbox__control-btn lightbox__control-btn--reset"
-            onClick={onReset}
-            type="button"
-            aria-label="Reset zoom"
-          >
-            <span aria-hidden="true">↺</span>
-            <span>Reset</span>
-          </button>
-          <button
-            className="lightbox__control-btn"
-            onClick={onZoomIn}
-            type="button"
-            aria-label="Zoom in"
-          >
-            +
-          </button>
-          <button
-            className="lightbox__control-btn lightbox__control-btn--close"
-            onClick={onClose}
-            type="button"
-            aria-label="Close preview"
-          >
-            ✕
-          </button>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-// ─── CropDrawer ───────────────────────────────────────────────────────────────
-
-interface CropDrawerProps {
-  rawImageUrl: string;
-  cropRect: { x: number; y: number; w: number; h: number };
-  imgRef: React.RefObject<HTMLImageElement | null>;
-  onImageLoad: () => void;
-  onDragStart: (e: React.MouseEvent) => void;
-  onResizeStart: (e: React.MouseEvent) => void;
-  onConfirm: () => void;
-  onCancel: () => void;
-}
-
-function CropDrawer({
-  rawImageUrl, cropRect, imgRef,
-  onImageLoad, onDragStart, onResizeStart,
-  onConfirm, onCancel,
-}: CropDrawerProps) {
-  return (
-    <div className="crop-drawer" role="dialog" aria-modal="true">
-      <div className="crop-drawer__inner">
-        <h3 className="crop-drawer__title">Crop &amp; Confirm</h3>
-
-        <div className="crop-drawer__stage">
-          <div className="crop-drawer__image-wrap">
-            <img
-              ref={imgRef}
-              src={rawImageUrl}
-              alt="Select crop area"
-              className="crop-drawer__image"
-              onLoad={onImageLoad}
-              draggable={false}
-            />
-            <div
-              className="crop-drawer__rect"
-              style={{
-                left: `${cropRect.x}px`,
-                top: `${cropRect.y}px`,
-                width: `${cropRect.w}px`,
-                height: `${cropRect.h}px`,
-              }}
-              onMouseDown={(e) => { e.preventDefault(); onDragStart(e); }}
-            >
-              <div
-                className="crop-drawer__handle"
-                onMouseDown={(e) => { e.stopPropagation(); onResizeStart(e); }}
-              />
-            </div>
-          </div>
-        </div>
-
-        <div className="crop-drawer__actions">
-          <button
-            className="btn btn--ghost crop-drawer__action-btn"
-            onClick={onCancel}
-            type="button"
-          >
-            <span aria-hidden="true">✕</span>
-            <span>Cancel</span>
-          </button>
-          <button
-            className="btn btn--primary crop-drawer__action-btn"
-            onClick={onConfirm}
-            type="button"
-          >
-            <span aria-hidden="true">✓</span>
-            <span>Crop &amp; Search</span>
-          </button>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-// ─── ImagePreview (pre-results, post-crop) ────────────────────────────────────
-
-interface ImagePreviewProps {
-  originalUrl: string;
-  croppedUrl: string | null;
-  searchLimit: number;
-  selectedCategories: string[];
-  loading: boolean;
-  onClear: () => void;
-  onRecrop: () => void;
-  onSetCategories: (cats: string[]) => void;
-  onSetLimit: (v: number) => void;
-  onSearch: () => void;
-}
-
-function ImagePreview({
-  originalUrl, croppedUrl, searchLimit,
-  selectedCategories, loading,
-  onClear, onRecrop, onSetCategories, onSetLimit, onSearch,
-}: ImagePreviewProps) {
-  return (
-    <div className="image-preview">
-      {/* Original */}
-      <div className="image-preview__pane">
-        <p className="image-preview__pane-title">Original Image</p>
-        <div className="image-preview__frame">
-          <img className="image-preview__img" src={originalUrl} alt="original" />
-          <div className="image-preview__frame-actions image-preview__frame-actions--top">
-            <button
-              className="btn btn--primary btn--sm"
-              onClick={onClear}
-              type="button"
-            >
-              <span aria-hidden="true">🗑️</span>
-              <span>Clear Search</span>
-            </button>
-          </div>
-          <div className="image-preview__frame-actions image-preview__frame-actions--bottom">
-            <button className="btn btn--outline btn--sm" onClick={onRecrop} type="button">
-              <span aria-hidden="true">✂️</span>
-              <span>Recrop</span>
-            </button>
-          </div>
-        </div>
-      </div>
-
-      {/* Cropped */}
-      {croppedUrl && (
-        <>
-          <div className="image-preview__pane">
-            <p className="image-preview__pane-title">Cropped Image</p>
-            <div className="image-preview__frame">
-              <img className="image-preview__img" src={croppedUrl} alt="cropped" />
-            </div>
-          </div>
-
-          <div className="image-preview__options">
-            <CategoryPicker selected={selectedCategories} onChange={onSetCategories} />
-          </div>
-
-          <div className="image-preview__search-row">
-            <LimitSlider value={searchLimit} onChange={onSetLimit} label="Limit" />
-            <button
-              className="btn btn--primary"
-              onClick={onSearch}
-              disabled={loading}
-              type="button"
-            >
-              <span aria-hidden="true">🔎</span>
-              <span>Search</span>
-            </button>
-          </div>
-        </>
-      )}
-    </div>
-  );
-}
-
-// ─── Hero ─────────────────────────────────────────────────────────────────────
-
-interface HeroProps {
-  textQuery: string;
-  setTextQuery: (v: string) => void;
-  onTextSearch: () => void;
-  onFileChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
-  fileInputId: string;
-  searchLimit: number;
-  onSetLimit: (v: number) => void;
-  selectedCategories: string[];
-  onSetCategories: (cats: string[]) => void;
-  loading: boolean;
-}
-
-function Hero({
-  textQuery, setTextQuery, onTextSearch,
-  onFileChange, fileInputId,
-  searchLimit, onSetLimit,
-  selectedCategories, onSetCategories,
-  loading,
-}: HeroProps) {
-  return (
-    <div className="hero">
-      <header className="hero__header">
-  <div className="hero__eyebrow">Fabric Intelligence</div>
-
-  <h1 className="hero__title">
-    Find the clothing
-    <br />
-    <span className="hero__title-accent">
-      you couldn't find.
-    </span>
-  </h1>
-
-  <p className="hero__subtitle">
-    Visual &amp; semantic search — powered by vectors
-  </p>
-</header>
-
-      <div className="hero__search-controls">
-        {/* Text search */}
-        <div className="hero__search-row">
-          <input
-            type="text"
-            className="hero__search-input"
-            placeholder="Search by name, fabric or color"
-            value={textQuery}
-            onChange={(e) => setTextQuery(e.target.value)}
-            onKeyDown={(e) => e.key === "Enter" && onTextSearch()}
-          />
-          <button
-            className="btn btn--primary"
-            onClick={onTextSearch}
-            disabled={loading || !textQuery.trim()}
-            type="button"
-          >
-            Search
-          </button>
-        </div>
-
-        {/* Limit */}
-        <LimitSlider value={searchLimit} onChange={onSetLimit} />
-
-        {/* Divider */}
-        <div className="hero__divider">or</div>
-
-        {/* Image upload */}
-        <input
-          id={fileInputId}
-          type="file"
-          accept="image/*"
-          onChange={onFileChange}
-          className="hero__file-input"
-        />
-        <button
-          className="btn btn--ghost hero__upload-btn"
-          onClick={() => document.getElementById(fileInputId)?.click()}
-          type="button"
-        >
-          <span aria-hidden="true">📷</span>
-          <span>Drop your Image</span>
-        </button>
-      </div>
-
-      {/* Category filter */}
-      <div className="hero__categories">
-        <CategoryPicker selected={selectedCategories} onChange={onSetCategories} />
-      </div>
-    </div>
-  );
-}
 
 // ─── Main Page ────────────────────────────────────────────────────────────────
 
@@ -1020,7 +85,7 @@ export default function Search() {
 
   // ── File input ─────────────────────────────────────────────────────────────
 
-  const onFileChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+  const onFileChange = (e: ChangeEvent<HTMLInputElement>) => {
     const f = e.target.files?.[0] ?? null;
     e.target.value = "";
     if (!f) return;
@@ -1223,15 +288,15 @@ export default function Search() {
     };
   }, [lightboxOpen, drawerOpen, rawImageUrl]);
 
-  const onLbWheel: React.WheelEventHandler<HTMLDivElement> = (e) => {
+  const onLbWheel: WheelEventHandler<HTMLDivElement> = (e) => {
     e.preventDefault();
     setLbScale((s) => Math.min(MAX_SCALE, Math.max(MIN_SCALE, s + (e.deltaY > 0 ? -ZOOM_STEP : ZOOM_STEP))));
   };
-  const onLbMouseDown: React.MouseEventHandler<HTMLDivElement> = (e) => {
+  const onLbMouseDown: MouseEventHandler<HTMLDivElement> = (e) => {
     draggingLbRef.current = true;
     lastLbPosRef.current = { x: e.clientX, y: e.clientY };
   };
-  const onLbMouseMove: React.MouseEventHandler<HTMLDivElement> = (e) => {
+  const onLbMouseMove: MouseEventHandler<HTMLDivElement> = (e) => {
     if (!draggingLbRef.current) return;
     const dx = e.clientX - lastLbPosRef.current.x;
     const dy = e.clientY - lastLbPosRef.current.y;
@@ -1414,7 +479,7 @@ export default function Search() {
 
         {/* Hero */}
         {showHero && (
-          <Hero
+          <SearchHero
             textQuery={textQuery}
             setTextQuery={setTextQuery}
             onTextSearch={handleTextSearch}
@@ -1459,7 +524,7 @@ export default function Search() {
             page={page}
             totalPages={totalPages}
             selectedCategories={selectedCategories}
-            onSetCategories={handleCategoryChange}   // ✅ FIXED
+            onSetCategories={handleCategoryChange}
             onPrev={safePrev}
             onNext={safeNext}
             onZoom={openLightbox}
